Tratamento de Exceptions Universo

+++ src/main/java/projeto/herois/controller/UniversosController.java
@@ -3,6 +3,10 @@ package projeto.herois.controller;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,6 +16,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import projeto.herois.model.Universos;
+import projeto.herois.payload.ApiResponse;
 import projeto.herois.repository.CrudUniversos;
 
 @RestController
@@ -22,13 +27,15 @@ public class UniversosController {
 	
 	@GetMapping("/universos")
 	public Iterable<Universos> getAllUniversos(){
-		return cu.findAll();
-		
+		return cu.findAll();	
 	}
 	
 	@GetMapping("/universo/{idUniverso}")
-	public Universos getUniversoById(@PathVariable("idUniverso") UUID idUniverso) {
+	public Universos getUniversoById(@PathVariable("idUniverso") UUID idUniverso) throws NotFoundException {
 		Universos universo = this.cu.findById(idUniverso).orElse(null);
+		if (universo == null) {
+			throw new NotFoundException();
+		}
 		return universo;
 	}
 	
@@ -38,13 +45,26 @@ public class UniversosController {
 	}
 
 	@DeleteMapping("/deletarUniverso/{idUniverso}")
-	public void deleteUniversoById(@PathVariable("idUniverso") UUID idUniverso) {
-		this.cu.deleteById(idUniverso);
+	public ResponseEntity<?> ResponseEntity (@PathVariable("idUniverso") UUID idUniverso) throws NotFoundException{
+		Universos universo = this.cu.findById(idUniverso).orElse(null); {
+		if (universo == null) {
+			throw new NotFoundException();
+		}
+		else {
+			this.cu.deleteById(idUniverso);
+			return new ResponseEntity<Object>(new ApiResponse(true, "Universo deletado!"),new HttpHeaders(), HttpStatus.OK);
+			}
+		}
 	}
 		
 	@PutMapping("/atualizarUniverso")
-	public Universos updateUniverso(@RequestBody Universos universo) {
-		return cu.save(universo);
-	}
-	
+	public ResponseEntity<?> updateUniverso(@RequestBody Universos universo) {
+		try {
+		getUniversoById(universo.getIdUniverso());
+		} catch (NotFoundException e) {
+		return new ResponseEntity<Object>(new ApiResponse(false, "Universo não encontrado!"),new HttpHeaders(), HttpStatus.NOT_FOUND);
+		}
+		cu.save(universo);
+		return new ResponseEntity<Object>(new ApiResponse(true, "Universo atualizado!"),new HttpHeaders(), HttpStatus.OK);
+		}
 }
